+++ frontend/src/App.js
@@ -1,13 +1,10 @@
 import React from "react";
-import NavBar from "./component/SharedComponents/NavBar";
-import Footer from "./component/SharedComponents/Footer";
+import Home from "./pages/Home";
 
 
 function App() {
   return (
     <div className="App">
-      <NavBar></NavBar>
-      <Footer></Footer>
     </div>
   )
 }
