+++ frontend/src/component/LandingPageComponents/HeroSection.jsx
@@ -5,16 +5,24 @@ import { motion } from "framer-motion";
 const HeroSection = () => {
 
   return (
-    <section class="relative font-sans bg-white">
-      
-        <div class="relative w-[100%] lg:absolute lg:right-0 lg:w-[70vw]">
+    <section class="font-sans bg-white ">
+      <div className=' top-0  lg:h-[80vh] xl:h-[95vh] 2xl:h-[100vh]'>
+        <div class=" w-[100%] lg:w-[80%]   ">
           <motion.img
             src={images.HeroPicture} 
             alt="Hero-Image"
             initial={{ x: "60%", y: "-60%", opacity: 0 }}
             animate={{ x: 0, y: 0, opacity: 1 }}
             transition={{ duration: 1, ease: "easeOut" }}
-            class={''}
+            class="hidden lg:block  lg:w-[75%] xl:w-[80%] lg:h-auto lg:max-h-[100vh] lg:right-0  lg:absolute"
+          />
+          <motion.img
+            src={images.HeroPictureMobileView} 
+            alt="Hero-Image"
+            initial={{ x: "60%", y: "-60%", opacity: 0 }}
+            animate={{ x: 0, y: 0, opacity: 1 }}
+            transition={{ duration: 1, ease: "easeOut" }}
+            class=" w-full lg:hidden"
           />
         </div>
       
@@ -22,25 +30,22 @@ const HeroSection = () => {
           initial={{ x: "-50%", opacity: 0 }}
           animate={{ x: 0, opacity: 1 }}
           transition={{ duration: 1, ease: "easeOut" }}
-          class="container relative mx-auto px-10 py-5 w-[100%] z-10 text-center lg-absolute lg:mx-14 lg:top-[48vh] lg:w-1/2 lg:text-left xl:top-[55vh]">
+          class="container lg:relative mx-auto  py-5 lg:py-0 w-[100%]   text-center z-50   lg:ml-[5rem] xl:ml-[6rem] 2xl:ml-[9rem] lg:top-[36vh] lg:w-[40%] lg:text-left xl:top-[50vh] 2xl:top-[60vh] ">
 
-            <h1 class="mb-6 text-[26px] text-gradient font-bold sm:text-[32px] md:text-4xl xl:text-5xl">
+            <h1 class="mb-6 text-[28px] text-gradient font-bold sm:text-4xl md:text-5xl lg:text-[2.5rem]  xl:text-5xl ">
               Unleashing <br></br> Innovation Through <br></br> Lines of Code
             </h1>
-            <p class="mb-6 text-[15px] text-[#7A7878] md:text-base">
+            <p class="mb-6 text-[15px] sm:text-[17px] md:text-[20px] text-[#7A7878]  lg:w-[70%]">
               We use the power of technology to push the boundaries in our digital age.
             </p>
-            <button class="px-6 py-2 mx-2 bg-gradient font-semibold text-[12px] text-white transition duration-600 rounded-[15px] shadow-md hover:bg-transparent md:py-3 md:px-8 lg:mx-0">
-              Discover More
-            </button>
-            <button class="bg-gradient font-semibold rounded-[15px] text-[10px] p-[3px] shadow-md md:p-[4px] md:ml-5 lg:invisible">
-              <span class="flex w-full bg-white text-[#1C3A7B] rounded-[14px] px-4 py-1 md:py-2 md:px-5">
-                Book a Free Consultation
-              </span>
-            </button>
+            <div className='flex flex-col  h-[6rem] sm:h-[8rem] lg:h-auto justify-between'>
+            <button className=' lg:flex font-sans font-bold text-[#ffffff] text-[14px] lg:text-[14px] xl:text-[16px] bg-gradient lg:h-[3rem] h-[2.5rem] sm:h-[3rem] xl:h-[3.5rem] w-full lg:w-[12rem] xl:w-[15rem]  items-center justify-center rounded-[20px] hover:text-gradient hover:border-4 hover:border-[#008A7D] transition-all duration-500 ease-linear '>Discover More</button>
+            <button className='font-sans font-bold lg:hidden text-[#008A7D] text-[14px]   h-[2.5rem] sm:h-[3rem] w-full   rounded-[20px] hover:text-[#ffff] hover:bg-gradient border-4 border-[#008A7D] hover:border-0 hover:bg- transition-all duration-500 ease-linear '>Book a Free Consultation</button>
+            </div>
+           
         
         </motion.div>
-      
+        </div>
     </section>
   )
 }
